requirements.txt
Flask==2.2.2
Flask_Cors==3.0.10
mysql_connector_repackaged==0.3.1